Remove theano.gof.unify-specific objects from theano.gof.utils
