+++ theano/gof/utils.py
@@ -461,118 +461,6 @@ def toposort(prereqs_d):
     return seq
 
 
-class Keyword:
-    def __init__(self, name, nonzero=True):
-        self.name = name
-        self.nonzero = nonzero
-
-    def __nonzero__(self):
-        # Python 2.x
-        return self.__bool__()
-
-    def __bool__(self):
-        # Python 3.x
-        return self.nonzero
-
-    def __str__(self):
-        return f"<{self.name}>"
-
-    def __repr__(self):
-        return f"<{self.name}>"
-
-
-ABORT = Keyword("ABORT", False)
-RETRY = Keyword("RETRY", False)
-FAILURE = Keyword("FAILURE", False)
-
-
-simple_types = (int, str, float, bool, type(None), Keyword)
-
-
-ANY_TYPE = Keyword("ANY_TYPE")
-FALL_THROUGH = Keyword("FALL_THROUGH")
-
-
-def comm_guard(type1, type2):
-    def wrap(f):
-        old_f = f.__globals__[f.__name__]
-
-        def new_f(arg1, arg2, *rest):
-            if (type1 is ANY_TYPE or isinstance(arg1, type1)) and (
-                type2 is ANY_TYPE or isinstance(arg2, type2)
-            ):
-                pass
-            elif (type1 is ANY_TYPE or isinstance(arg2, type1)) and (
-                type2 is ANY_TYPE or isinstance(arg1, type2)
-            ):
-                arg1, arg2 = arg2, arg1
-            else:
-                return old_f(arg1, arg2, *rest)
-
-            variable = f(arg1, arg2, *rest)
-            if variable is FALL_THROUGH:
-                return old_f(arg1, arg2, *rest)
-            else:
-                return variable
-
-        new_f.__name__ = f.__name__
-
-        def typename(type):
-            if isinstance(type, Keyword):
-                return str(type)
-            elif isinstance(type, (tuple, list)):
-                return "(" + ", ".join([x.__name__ for x in type]) + ")"
-            else:
-                return type.__name__
-
-        new_f.__doc__ = (
-            str(old_f.__doc__)
-            + "\n"
-            + ", ".join([typename(type) for type in (type1, type2)])
-            + "\n"
-            + str(f.__doc__ or "")
-        )
-        return new_f
-
-    return wrap
-
-
-def type_guard(type1):
-    def wrap(f):
-        old_f = f.__globals__[f.__name__]
-
-        def new_f(arg1, *rest):
-            if type1 is ANY_TYPE or isinstance(arg1, type1):
-                variable = f(arg1, *rest)
-                if variable is FALL_THROUGH:
-                    return old_f(arg1, *rest)
-                else:
-                    return variable
-            else:
-                return old_f(arg1, *rest)
-
-        new_f.__name__ = f.__name__
-
-        def typename(type):
-            if isinstance(type, Keyword):
-                return str(type)
-            elif isinstance(type, (tuple, list)):
-                return "(" + ", ".join([x.__name__ for x in type]) + ")"
-            else:
-                return type.__name__
-
-        new_f.__doc__ = (
-            str(old_f.__doc__)
-            + "\n"
-            + ", ".join([typename(type) for type in (type1,)])
-            + "\n"
-            + str(f.__doc__ or "")
-        )
-        return new_f
-
-    return wrap
-
-
 def flatten(a):
     """
     Recursively flatten tuple, list and set in a list.
